refactor: added debouncing of input validations
